+++ tools/dart-code-gen/lib/src/field_validator_factory.dart
@@ -20,10 +20,10 @@
 
 import 'package:code_builder/code_builder.dart';
 import 'package:dart_code_gen/google/protobuf/descriptor.pb.dart';
-import 'package:dart_code_gen/spine/options.pb.dart';
 
 import 'constraint_violation.dart';
-import 'imports.dart';
+import 'float_validator_factory.dart';
+import 'string_validator_factory.dart';
 import 'validator_factory.dart';
 
 /// Factory of validation code for a given message field.
@@ -36,16 +36,16 @@ class FieldValidatorFactory {
     /// The field to validate.
     final FieldDescriptorProto field;
 
-    FieldValidatorFactory._(this.validatorFactory, this.field);
+    FieldValidatorFactory(this.validatorFactory, this.field);
 
     /// Creates a new `FieldValidatorFactory` for the given field.
     factory FieldValidatorFactory.forField(FieldDescriptorProto field, ValidatorFactory factory) {
         var type = field.type;
         switch (type) {
             case FieldDescriptorProto_Type.TYPE_STRING:
-                return new _StringValidatorFactory(factory, field);
+                return new StringValidatorFactory(factory, field);
             case FieldDescriptorProto_Type.TYPE_FLOAT:
-                return new _FloatValidatorFactory(factory, field);
+                return new FloatValidatorFactory(factory, field);
         }
         return null;
     }
@@ -58,8 +58,8 @@ class FieldValidatorFactory {
     /// the [ValidatorFactory.violationList] of the [validatorFactory].
     ///
     Code createFieldValidator(Expression fieldValue) {
-        var statements = _rules()
-            .map((r) => r.eval(fieldValue))
+        var statements = rules()
+            .map((r) => r._eval(fieldValue))
             .map((expression) => expression.statement)
             .toList();
         return statements.isNotEmpty
@@ -68,15 +68,15 @@ class FieldValidatorFactory {
     }
 
     /// Obtains validation rules to apply to the field..
-    Iterable<_Rule> _rules() => null;
+    Iterable<Rule> rules() => null;
 
-    _Rule _rule(LazyCondition condition, LazyViolation violation) {
-        return _Rule(condition, violation, validatorFactory.violationList);
+    Rule newRule(LazyCondition condition, LazyViolation violation) {
+        return Rule._(condition, violation, validatorFactory.violationList);
     }
 
     /// Generates an expression which constructs a `ConstraintViolation` for a missing required
     /// field.
-    Expression _requiredMissing() {
+    Expression requiredMissing() {
         return violationRef.call([literalString('Field must be set.'),
                                   literalString(validatorFactory.fullTypeName),
                                   literalList([field.name])]);
@@ -89,7 +89,7 @@ class FieldValidatorFactory {
 ///
 /// Typically, one rule represents one validation option for a given field type.
 ///
-class _Rule {
+class Rule {
 
     final LazyCondition _condition;
     final LazyViolation _violation;
@@ -106,7 +106,7 @@ class _Rule {
     /// the `ConstraintViolation`s. If the rule is broken, the violation generated by [_violation]
     /// is added to the list.
     ///
-    _Rule(this._condition, this._violation, this._targetViolationList);
+    Rule._(this._condition, this._violation, this._targetViolationList);
 
     /// Produces a ternary operator which creates a new violation if the string is empty.
     ///
@@ -119,7 +119,7 @@ class _Rule {
     /// `code_builder` does not support `if` statements, so a ternary conditional operator has to
     /// be used.
     ///
-    Expression eval(Expression fieldValue) {
+    Expression _eval(Expression fieldValue) {
         var ternaryOperator = _condition(fieldValue).conditional(
             _targetViolationList
                 .property('add')
@@ -134,109 +134,3 @@ typedef Expression LazyViolation(Expression fieldValue);
 
 /// A function of a field value expresion to a boolean expression representing a constraint.
 typedef Expression LazyCondition(Expression fieldValue);
-
-/// A [FieldValidatorFactory] for `string` fields.
-///
-/// Currently, the only supported option is `(required)`.
-///
-class _StringValidatorFactory extends FieldValidatorFactory {
-
-    _StringValidatorFactory(ValidatorFactory validatorFactory, FieldDescriptorProto field)
-        : super._(validatorFactory, field);
-
-    Iterable<_Rule> _rules() {
-        var options = field.options;
-        var rules = <_Rule>[];
-        if (options.hasExtension(Options.required) && options.getExtension(Options.required)) {
-            var requiredString = _rule((v) => v.property('isEmpty'),
-                                       (v) => _requiredMissing());
-            rules.add(requiredString);
-        }
-        if (options.hasExtension(Options.pattern)) {
-            PatternOption pattern = options.getExtension(Options.pattern);
-            var rule = _rule((v) => refer('RegEx').newInstance([literalString(pattern.regex)])
-                                                  .property('stringMatch')
-                                                  .call([v])
-                                                  .equalTo(v),
-                             (v) => _patternMismatch(pattern.regex));
-            rules.add(rule);
-        }
-        return rules;
-    }
-
-    Expression _patternMismatch(String pattern) {
-        var message = 'String must match the regular expression `$pattern`';
-        return violationRef.call([literalString(message),
-                                  literalString(validatorFactory.fullTypeName),
-                                  literalList([field.name])]);
-    }
-}
-
-/// A [FieldValidatorFactory] for `float` fields.
-///
-/// Currently, the supported options are `(min)` and `(max)`.
-///
-class _FloatValidatorFactory extends FieldValidatorFactory {
-
-    _FloatValidatorFactory(ValidatorFactory validatorFactory, FieldDescriptorProto field)
-        : super._(validatorFactory, field);
-
-    Iterable<_Rule> _rules() {
-        var rules = <_Rule>[];
-        var options = field.options;
-        if (options.hasExtension(Options.min)) {
-            _Rule min = _minRule(options);
-            rules.add(min);
-        }
-        if (options.hasExtension(Options.max)) {
-            _Rule max = _maxRule(options);
-            rules.add(max);
-        }
-        return rules;
-    }
-
-    _Rule _minRule(FieldOptions options) {
-      var min = options.getExtension(Options.min) as MinOption;
-      var bound = double.parse(min.value);
-      var exclusive = min.exclusive;
-      var literal = literalNum(bound);
-      var check = exclusive
-                  ? (Expression v) => v.lessOrEqualTo(literal)
-                  : (Expression v) => v.lessThan(literal);
-      var requiredString = _rule((v) => check(v), _outOfBound);
-      return requiredString;
-    }
-
-    _Rule _maxRule(FieldOptions options) {
-      var max = options.getExtension(Options.max) as MaxOption;
-      var bound = double.parse(max.value);
-      var exclusive = max.exclusive;
-      var literal = literalNum(bound);
-      var check = exclusive
-                  ? (Expression v) => v.greaterOrEqualTo(literal)
-                  : (Expression v) => v.greaterThan(literal);
-      var requiredString = _rule((v) => check(v), _outOfBound);
-      return requiredString;
-    }
-
-    // TODO:2019-10-14:dmytro.dashenkov: Support custom error messages based on the option value.
-    // https://github.com/SpineEventEngine/base/issues/482
-    Expression _outOfBound(Expression value) {
-        var param = 'v';
-        var standardPackage = validatorFactory.properties.standardPackage;
-        var floatValue = refer('FloatValue', protoWrappersImport(standardPackage))
-            .newInstance([])
-            .property('copyWith')
-            .call([Method((b) => b
-            ..requiredParameters.add(Parameter((b) => b.name = param))
-            ..body = refer(param)
-                .property('value')
-                .assign(value)
-                .statement).closure]);
-        var any = refer('Any', protoAnyImport(standardPackage)).property('pack').call([floatValue]);
-        return violationRef.call([literalString('Float field is out of bound.'),
-                                  literalString(validatorFactory.fullTypeName),
-                                  literalList([field.name]),
-                                  any]);
-    }
-}
