+++ tools/dart-code-gen/lib/src/string_validator_factory.dart
@@ -1,0 +1,80 @@
+/*
+ * Copyright 2019, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import 'package:code_builder/code_builder.dart';
+import 'package:dart_code_gen/google/protobuf/descriptor.pb.dart';
+import 'package:dart_code_gen/spine/options.pb.dart';
+
+import 'constraint_violation.dart';
+import 'field_validator_factory.dart';
+import 'validator_factory.dart';
+
+/// A [FieldValidatorFactory] for `string` fields.
+///
+/// Currently, the only supported option is `(required)`.
+///
+class StringValidatorFactory extends FieldValidatorFactory {
+
+    StringValidatorFactory(ValidatorFactory validatorFactory, FieldDescriptorProto field)
+        : super(validatorFactory, field);
+
+    Iterable<Rule> rules() {
+        var options = field.options;
+        var rules = <Rule>[];
+        if (options.hasExtension(Options.required) && options.getExtension(Options.required)) {
+            Rule requiredString = _requiredRule();
+            rules.add(requiredString);
+        }
+        if (options.hasExtension(Options.pattern)) {
+            Rule rule = _patternRule(options);
+            rules.add(rule);
+        }
+        return rules;
+    }
+
+    Rule _requiredRule() {
+        var requiredString = newRule((v) => v.property('isEmpty'),
+                                  (v) => requiredMissing());
+        return requiredString;
+    }
+
+    /// Creates a validation rule which matches a string upon a regular expression.
+    ///
+    /// The whole string must match the regex. To check that, the generated code applies the `RegEx`
+    /// to the string and compares the first match to the value of the initial string. If the values
+    /// are identical, the check passes.
+    ///
+    Rule _patternRule(FieldOptions options) {
+        PatternOption pattern = options.getExtension(Options.pattern);
+        var rule = newRule((v) => refer('RegEx').newInstance([literalString(pattern.regex)])
+                                                .property('stringMatch')
+                                                .call([v])
+                                                .equalTo(v),
+                           (v) => _patternMismatch(pattern.regex));
+        return rule;
+    }
+
+    Expression _patternMismatch(String pattern) {
+        var message = 'String must match the regular expression `$pattern`';
+        return violationRef.call([literalString(message),
+                                     literalString(validatorFactory.fullTypeName),
+                                     literalList([field.name])]);
+    }
+}
